Fix MongoDB Sink test
Related to spring-cloud-stream-app-starters/app-starters-release#168

Starting Spring Data MongoDB 2.1, there is no more copy/paste for
provided `Map` to a new instance and `MongoDbTemplate` tries to insert
an `id` to the provided `Map`.
Therefore we can't use `Collections.singletonMap()` as an entity to save.
Also for optimization reason the `_class` field is not populated on the
document for `Map` entities any more.
Use `@AutoConfigureDataMongo` instead of `spring.data.mongodb.port=0`
property